testes em andamento

+++ src/Api.Domain/Seguranca/TokenConfigurations.cs
@@ -8,6 +8,6 @@ namespace Domain.Security
     {
         public string Audience { get; set; }
         public string Issuer { get; set; }
-        public int Seconds { get; set; }
+        public double Seconds { get; set; }
     }
 }
